FrameCacheManagerを複数のファイルに分割
