+++ src/Beutl/Models/FrameCacheManager.CacheEntry.cs
@@ -1,0 +1,104 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Beutl.Media;
+using Beutl.Media.Pixel;
+using Beutl.Media.Source;
+
+using OpenCvSharp;
+
+namespace Beutl.Models;
+
+public partial class FrameCacheManager
+{
+    private class CacheEntry : IDisposable
+    {
+        private Mat _mat;
+
+        public CacheEntry(Ref<Bitmap<Bgra8888>> bitmap, FrameCacheOptions options)
+        {
+            SetBitmap(bitmap, options);
+            using (Ref<Bitmap<Bgra8888>> t = bitmap.Clone())
+            {
+                _mat = ToYUV(t.Value, options);
+            }
+
+            LastAccessTime = DateTime.UtcNow;
+        }
+
+        public DateTime LastAccessTime { get; private set; }
+
+        public int ByteCount => (int)(_mat.DataEnd - _mat.DataStart);
+
+        public bool IsLocked { get; set; }
+
+        public void SetBitmap(Ref<Bitmap<Bgra8888>> bitmap, FrameCacheOptions options)
+        {
+            _mat?.Dispose();
+            using (Ref<Bitmap<Bgra8888>> t = bitmap.Clone())
+            {
+                _mat = ToYUV(t.Value, options);
+            }
+
+            LastAccessTime = DateTime.UtcNow;
+        }
+
+        public Ref<Bitmap<Bgra8888>> GetBitmap()
+        {
+            LastAccessTime = DateTime.UtcNow;
+            return Ref<Bitmap<Bgra8888>>.Create(ToBitmap(_mat));
+        }
+
+        private static Mat ToYUV(Bitmap<Bgra8888> bitmap, FrameCacheOptions options)
+        {
+            var result = new Mat(bitmap.Height, bitmap.Width, MatType.CV_8UC4, bitmap.Data);
+
+            PixelSize size = new(bitmap.Width, bitmap.Height);
+            PixelSize newSize = options.GetSize(size);
+            if (newSize != size)
+            {
+                Mat tmp = result.Resize(new Size(newSize.Width, newSize.Height));
+                result.Dispose();
+                result = tmp;
+            }
+
+            if (options.ColorType == FrameCacheColorType.YUV)
+            {
+                var mat = new Mat((int)(result.Rows * 1.5), result.Cols, MatType.CV_8UC1);
+                Cv2.CvtColor(result, mat, ColorConversionCodes.BGRA2YUV_I420);
+                result.Dispose();
+                result = mat;
+            }
+
+            return result;
+        }
+
+        private static unsafe Bitmap<Bgra8888> ToBitmap(Mat mat)
+        {
+            Bitmap<Bgra8888>? bitmap;
+            if (mat.Type() == MatType.CV_8UC4)
+            {
+                bitmap = new Bitmap<Bgra8888>(mat.Width, mat.Height);
+                Buffer.MemoryCopy((void*)mat.Data, (void*)bitmap.Data, bitmap.ByteCount, bitmap.ByteCount);
+            }
+            else
+            {
+                bitmap = new Bitmap<Bgra8888>(mat.Width, (int)(mat.Height / 1.5));
+                using var bgra = new Mat(bitmap.Height, bitmap.Width, MatType.CV_8UC4, bitmap.Data);
+
+                Cv2.CvtColor(mat, bgra, ColorConversionCodes.YUV2BGRA_I420);
+            }
+
+            return bitmap;
+        }
+
+        public void Dispose()
+        {
+            _mat?.Dispose();
+            _mat = null!;
+        }
+    }
+}
